[GPU] transformations pipeline is extracted to separate class
